add generic content view usable in apod view and photo detail view
